🎨 split component into separate file
