+++ src/app/(main)/accounts/page.tsx
@@ -1,19 +1,20 @@
-import { PurchaseHistoryItemWrap } from '@components/product/ProductItem';
+import OrderHistoryPage from "@app/(main)/accounts/purchase/page";
+
 
 export default function Account() {
   return (
     <>
-      {/* 구매내역이 디폴트 */}
+      {/* 구매내역 탭이 디폴트 */}
       <PurchaseHistoryTab />
     </>
   );
 }
 
-// 구매내역 탭
+// 구매내역 탭 밑에 리스트
 function PurchaseHistoryTab() {
   return (
     <>
-      <PurchaseHistoryItemWrap /> 
+      <OrderHistoryPage />
     </>
   );
 }
