+++ src/app/(main)/accounts/purchase/page.tsx
@@ -1,0 +1,9 @@
+import PurchaseHistoryItemWrap from "@app/(main)/accounts/purchase/OrderHistory";
+
+export default function OrderHistoryPage(){
+  return(
+    <>
+       <PurchaseHistoryItemWrap /> 
+    </>
+  );
+}
